WIP Neon vectorization

+++ protocol/mpdu.h
@@ -11,7 +11,7 @@
 #define MPDU_PER_PERIOD (3*MPDU_PER_LINE + 1)
 #define MCU_PER_LINE (MCU_PER_MPDU * MPDU_PER_LINE)
 #define MPDU_US_PER_LINE (1220*1000) /* Imprecise, lower bound only */
-#define US_PER_DAY (1000L * 1000 * 86400)
+#define US_PER_DAY ((uint64_t)1000L * 1000L * 86400L)
 
 typedef struct {
 	uint8_t day[2];
